+++ Notepad2e.vcxproj
@@ -324,7 +324,6 @@
     <ClCompile Include="src\Edit.c" />
     <ClCompile Include="src\Extension\EditHelper.cpp" />
     <ClCompile Include="src\Extension\HLSelection.c" />
-    <ClCompile Include="src\Extension\hl_helper.cpp" />
     <ClCompile Include="src\Extension\InlineProgressBarCtrl.c" />
     <ClCompile Include="src\Extension\StrToHex.c" />
     <ClCompile Include="src\Extension\tinyexpr\tinyexpr.c" />
@@ -398,6 +397,7 @@
     <ClInclude Include="src\Dialogs.h" />
     <ClInclude Include="src\Dlapi.h" />
     <ClInclude Include="src\Edit.h" />
+    <ClInclude Include="src\Extension\EditHelper.h" />
     <ClInclude Include="src\Extension\HLSelection.h" />
     <ClInclude Include="src\Extension\InlineProgressBarCtrl.h" />
     <ClInclude Include="src\Extension\SciCall.h" />
