+++ src/Edit.c
@@ -34,6 +34,7 @@
 #include "dialogs.h"
 #include "helpers.h"
 #include "resource.h"
+#include "EditHelper.h"
 #include "HLSelection.h"
 #include "StrToHex.h"
 
@@ -77,9 +78,6 @@ extern	int       iEncoding;
 // Supported Encodings
 WCHAR wchANSI[8] = L"";
 WCHAR wchOEM[8] = L"";
-
-WCHAR	hl_last_html_tag[0xff] = L"<tag>";
-WCHAR	hl_last_html_end_tag[0xff] = L"</tag>";
 
 NP2ENCODING mEncoding[] = {
   { NCP_DEFAULT | NCP_RECODE, 0, "ansi,ansi,ascii,", 61000, L"" },
@@ -159,9 +157,6 @@ NP2ENCODING mEncoding[] = {
 
 extern LPMRULIST mruFind;
 extern LPMRULIST mruReplace;
-
-#define HL_IS_LITERAL(CH) hl_iswordchar(CH)
-#define HL_IS_SPACE(CH) hl_isspace(CH)
 
 //=============================================================================
 //
@@ -2031,84 +2026,6 @@ void EditUnescapeCChars(HWND hwnd)
       MsgBox(MBINFO, IDS_SELRECT);
     }
   }
-}
-
-/*
- * HL STRIP HTML TAGS
- **/
-void HL_Strip_html_tags(HWND hwnd)
-{
-  struct Sci_TextToFind ttf1, ttf2;
-  int selbeg, selend, res, len;
-
-  //
-  selbeg = SendMessage(hwnd, SCI_GETSELECTIONSTART, 0, 0);
-  selend = SendMessage(hwnd, SCI_GETSELECTIONEND, 0, 0);
-  len = SendMessage(hwnd, SCI_GETTEXTLENGTH, 0, 0);
-  SendMessage(hwnd, SCI_BEGINUNDOACTION, 0, 0);
-  //
-  if (0 == selend - selbeg)
-  {
-    ttf1.chrg.cpMin = selbeg;
-    ttf1.chrg.cpMax = 0;
-    ttf1.lpstrText = "<";
-    res = SendMessage(hwnd, SCI_FINDTEXT, 0, (LPARAM)&ttf1);
-    if (-1 != res)
-    {
-      ttf2.chrg.cpMin = ttf1.chrgText.cpMax;
-      ttf2.chrg.cpMax = len;
-      ttf2.lpstrText = ">";
-      res = SendMessage(hwnd, SCI_FINDTEXT, 0, (LPARAM)&ttf2);
-      if (-1 != res)
-      {
-        SendMessage(hwnd, SCI_DELETERANGE, ttf1.chrgText.cpMin, ttf2.chrgText.cpMax - ttf1.chrgText.cpMin);
-        SendMessage(hwnd, SCI_SETSEL, ttf1.chrgText.cpMin, ttf1.chrgText.cpMin);
-      }
-    }
-  }
-  else
-  {
-    while (1)
-    {
-      ttf1.chrg.cpMin = selbeg;
-      ttf1.chrg.cpMax = selend;
-      ttf1.lpstrText = "<";
-      res = SendMessage(hwnd, SCI_FINDTEXT, 0, (LPARAM)&ttf1);
-      if (-1 != res)
-      {
-        ttf2.chrg.cpMin = ttf1.chrgText.cpMax;
-        ttf2.chrg.cpMax = selend;
-        ttf2.lpstrText = ">";
-        res = SendMessage(hwnd, SCI_FINDTEXT, 0, (LPARAM)&ttf2);
-        if (-1 != res)
-        {
-          int dlen = ttf2.chrgText.cpMax - ttf1.chrgText.cpMin;
-          SendMessage(hwnd, SCI_DELETERANGE, ttf1.chrgText.cpMin, dlen);
-          selend -= dlen;
-        }
-        else
-        {
-          break;
-        }
-      }
-      else
-      {
-        break;
-      }
-    }
-    SendMessage(hwnd, SCI_SETSEL, selbeg, selend);
-  }
-  SendMessage(hwnd, SCI_ENDUNDOACTION, 0, 0);
-}
-
-BOOL IsSelectionModeValid(HWND hwnd)
-{
-  if (SC_SEL_RECTANGLE == SendMessage(hwnd, SCI_GETSELECTIONMODE, 0, 0))
-  {
-    MsgBox(MBINFO, IDS_SELRECT);
-    return FALSE;
-  }
-  return TRUE;
 }
 
 //=============================================================================
